Rename namespaces and move files for better organization
Renamed the namespaces to match new directory structure and relocated the affected files. This improves the organization and readability of the project by categorizing related classes under more appropriate namespaces.
